Renamed file for init configs
